Refactor parse constructors.
