+++ sources/mandatory/parser/exec.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/09 14:28:45 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/09/11 20:10:26 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/09/11 21:10:56 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,8 @@ t_parse	*new_exec(char **tokens)
 {
 	t_parse	*new;
 
+	if (ft_strarr_len(tokens) == 0)
+		return (NULL);
 	new = new_parse();
 	new->type = PT_EXEC;
 	new->exec.cmd = ft_strdup(tokens[0]);
@@ -41,4 +43,22 @@ t_parse	*new_exec_length(char **tokens, int length)
 		length--;
 	}
 	return (new);
+}
+
+void	add_exec(t_parse_list **list, char **tokens)
+{
+	t_parse	*_parse;
+	char	**end;
+	int		length;
+
+	if (tokens == NULL || *tokens == NULL)
+		return ;
+	end = find_next_pipe(tokens);
+	length = end - tokens;
+	if (length <= 0)
+		return ;
+	_parse = new_exec_length(tokens, length);
+	if (_parse == NULL)
+		return ;
+	ft_dlst_add(list, _parse);
 }
